Memory leak / cleanup
- using weak self in getFollowers network manager function in FollowerListVC to remove memory leak
- Added UIHelper struct and moved createThreeColumnFlowLayout function to it
